more cleanup of core things
