+++ modules/api/src/main/scala/vinyldns/api/repository/DataStoreLoader.scala
@@ -119,3 +119,5 @@ object DataStoreLoader {
     accessor.toEither.leftMap(errors => DataStoreStartupError(errors.toList.mkString(", ")))
   }
 }
+
+case class DataStoreStartupError(msg: String) extends Throwable(msg)
