+++ modules/core/src/main/scala/vinyldns/api/domain/membership/ListUsersResults.scala
@@ -16,20 +16,4 @@
 
 package vinyldns.api.domain.membership
 
-import cats.effect._
-import vinyldns.api.repository.Repository
-
-trait UserRepository extends Repository {
-
-  /*Looks up a user.  If the user is not found, or if the user's status is Deleted, will return None */
-  def getUser(userId: String): IO[Option[User]]
-
-  def getUsers(
-      userIds: Set[String],
-      exclusiveStartKey: Option[String],
-      pageSize: Option[Int]): IO[ListUsersResults]
-
-  def getUserByAccessKey(accessKey: String): IO[Option[User]]
-
-  def save(user: User): IO[User]
-}
+final case class ListUsersResults(users: Seq[User], lastEvaluatedId: Option[String])
