+++ modules/core/src/main/scala/vinyldns/api/domain/record/ListRecordSetChangesResults.scala
@@ -14,22 +14,10 @@
  * limitations under the License.
  */
 
-package vinyldns.api.domain.membership
+package vinyldns.api.domain.record
 
-import cats.effect._
-import vinyldns.api.repository.Repository
-
-trait UserRepository extends Repository {
-
-  /*Looks up a user.  If the user is not found, or if the user's status is Deleted, will return None */
-  def getUser(userId: String): IO[Option[User]]
-
-  def getUsers(
-      userIds: Set[String],
-      exclusiveStartKey: Option[String],
-      pageSize: Option[Int]): IO[ListUsersResults]
-
-  def getUserByAccessKey(accessKey: String): IO[Option[User]]
-
-  def save(user: User): IO[User]
-}
+case class ListRecordSetChangesResults(
+    items: List[RecordSetChange] = List[RecordSetChange](),
+    nextId: Option[String] = None,
+    startFrom: Option[String] = None,
+    maxItems: Int = 100)
