+++ modules/core/src/main/scala/vinyldns/api/domain/record/RecordSet.scala
@@ -19,74 +19,16 @@ package vinyldns.api.domain.record
 import java.util.UUID
 
 import org.joda.time.DateTime
-import vinyldns.api.domain.DomainHelpers.ensureTrailingDot
 
 object RecordType extends Enumeration {
   type RecordType = Value
   val A, AAAA, CNAME, PTR, MX, NS, SOA, SRV, TXT, SSHFP, SPF, UNKNOWN = Value
 }
+
 object RecordSetStatus extends Enumeration {
   type RecordSetStatus = Value
   val Active, Inactive, Pending, PendingUpdate, PendingDelete = Value
 }
-
-sealed trait RecordData
-
-final case class AData(address: String) extends RecordData
-
-final case class AAAAData(address: String) extends RecordData
-
-final case class CNAMEData(cname: String) extends RecordData
-
-object CNAMEData {
-  def apply(cname: String): CNAMEData =
-    new CNAMEData(ensureTrailingDot(cname))
-}
-
-final case class MXData(preference: Integer, exchange: String) extends RecordData
-
-object MXData {
-  def apply(preference: Integer, exchange: String): MXData =
-    new MXData(preference, ensureTrailingDot(exchange))
-}
-
-final case class NSData(nsdname: String) extends RecordData
-
-object NSData {
-  def apply(nsdname: String): NSData =
-    new NSData(ensureTrailingDot(nsdname))
-}
-
-final case class PTRData(ptrdname: String) extends RecordData
-
-object PTRData {
-  def apply(ptrdname: String): PTRData =
-    new PTRData(ensureTrailingDot(ptrdname))
-}
-
-final case class SOAData(
-    mname: String,
-    rname: String,
-    serial: Long,
-    refresh: Long,
-    retry: Long,
-    expire: Long,
-    minimum: Long)
-    extends RecordData
-
-final case class SPFData(text: String) extends RecordData
-
-final case class SRVData(priority: Integer, weight: Integer, port: Integer, target: String)
-    extends RecordData
-
-object SRVData {
-  def apply(priority: Integer, weight: Integer, port: Integer, target: String): SRVData =
-    new SRVData(priority, weight, port, ensureTrailingDot(target))
-}
-
-final case class SSHFPData(algorithm: Integer, typ: Integer, fingerprint: String) extends RecordData
-
-final case class TXTData(text: String) extends RecordData
 
 import RecordSetStatus._
 import RecordType._
